EnemyAI Rudimentary Stealth Extension
• Enemy has states
- idle
- chasing
- searching
- investigating
- shooting

the last 3 have not been implemented fully, but, effectively when the enemy sees your or is shot by you, it goes into chase mode. When you flee out of its range and then out of its site, it goes to the last place it saw you and enters searching mode (this will eventually make the enemy kind of search the area you were last seen in).

Finally, there is investigate (if you were to shoot out a light, drop a heavy object, or make a loud sound), where the enemy would go to where they heard the sound and then start searching.

Shooting is straight forward.