Light Detection - Player Side - Basic Implementation
Lights have sphere triggers at the ground, with a radius. If player goes in radius a raycast fires to them. Then the program determines how far from the center of the sphere the player is, AND if they are being blocked at all. It then sends this data to the player and updates the visibility. Then the gameManager gets ahold of it and displays the number on the bottom. Additionally, crouching and sprinting both modify the visibility by -15 for crouching and +15 for sprinting.

Next would be to implement this into the enemy detection system. I can do, unless someone else wants to.